% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AddParSettingFAST}
\alias{AddParSettingFAST}
\title{Add FAST model settings for a PRECASTObj object}
\usage{
AddParSettingFAST(PRECASTObj, ...)
}
\arguments{
\item{PRECASTObj}{a PRECASTObj object created by \code{\link[PRECAST]{CreatePRECASTObject}}.}

\item{...}{other arguments to be passed to \code{\link{model_set_FAST}} function.}
}
\value{
Return a revised PRECASTObj object with slot \code{parameterList} changed.
}
\description{
Add FAST model settings for a PRECASTObj object
}
\references{
None
}
