% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AddParSettingProFAST}
\alias{AddParSettingProFAST}
\title{Add ProFAST model settings for a PRECASTObj object}
\usage{
AddParSettingProFAST(PRECASTObj, ...)
}
\arguments{
\item{PRECASTObj}{a PRECASTObj object created by \code{\link{CreatePRECASTObject}}.}

\item{...}{other arguments to be passed to \code{\link{model_set_ProFAST}} function.}
}
\value{
Return a revised PRECASTObj object with slot \code{parameterList} changed.
}
\description{
Add ProFAST model settings for a PRECASTObj object
}
\references{
None
}
