% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Axrat_Filter.R
\name{Axrat_Filter}
\alias{Axrat_Filter}
\title{Axrat_filter}
\usage{
Axrat_Filter(
  RA_DEC,
  axrat_value = 0.35,
  savepassthru = FALSE,
  filtered_data = NULL,
  colours = c("g", "r", "i")
)
}
\arguments{
\item{RA_DEC}{String; Celestial Right Ascension and Declination of Input Frame separated by underscore (RA_Dec).}

\item{axrat_value}{Numeric scalar; Filter Cutoff Value. Sources with an axial ratio lower than the value are assumed to not be asteroids and are removed.}

\item{savepassthru}{Logical; should intermediate files be saved to directory? Can greatly increase size on disk but useful to see which objects are being filtered out.}

\item{filtered_data}{List; Flux filtered astronometric data. If not supplied will look for appropriate file in working directory.}

\item{colours}{Character vector; List containing detection bands for input fields. ProFAst will default to looking for g,r & i bands unless told otherwise.}
}
\value{
Data frame containing all axrat filtered sources.
}
\description{
Filter sources in data frame based on axial ratio value.
}
