\name{CosMx_subset}  
\alias{CosMx_subset}  
\title{A Seurat object including spatial transcriptomics dataset from CosMx platform}  
\description{  
  This data is a subset of  SCLC CosMx spatial transcriptomics  dataset.
}  
\usage{  
  data(CosMx_subset)  
}  
\format{  
  A Seurat object, including count matrix, sptial coordinates, and manual annotation.
}  
\source{  
  The data is from the CosMx SRT sequencing platform.
}  
\references{  
  None
}  
\examples{  
  # Show some examples of how to use the dataset.  
  data(CosMx_subset)  
  library(Seurat)
  CosMx_subset
}  
\keyword{datasets}
