% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{FAST}
\alias{FAST}
\title{Run FAST model for a PRECASTObj object}
\usage{
FAST(PRECASTObj, q = 15, fit.model = c("poisson", "gaussian"))
}
\arguments{
\item{PRECASTObj}{a PRECASTObj object created by \code{\link[PRECAST]{CreatePRECASTObject}}.}

\item{q}{an optional integer, specify the number of low-dimensional embeddings to extract in FAST}

\item{fit.model}{an optional string, specify the version of FAST to be fitted. The Gaussian version models the log-count matrices while the Poisson verions models the count matrices; default as poisson.}
}
\value{
Return a revised PRECASTObj object with slot \code{PRECASTObj@resList} added by a \code{FAST} compoonent.
}
\description{
Run FAST model for a PRECASTObj object
}
\references{
None
}
