% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flux_Filter.R
\name{Flux_Filter}
\alias{Flux_Filter}
\title{Flux_filter}
\usage{
Flux_Filter(RA_DEC, flux_value = 1, edge_buffer = 0.001, savepassthru = FALSE)
}
\arguments{
\item{RA_DEC}{Celestial Right Ascension and Declination of Input Frame separated by underscore (RA_Dec).}

\item{flux_value}{Numeric scalar; Flux filter cutoff value. Determined by dividing query filter flux by sum of fluxes in other bands , i.e. (g/r+i). Sources with an flux ratio lower than the value are assumed to not be asteroids and are removed.}

\item{edge_buffer}{Numeric scalar; Edge boundary value within which positive hits are ignored. Useful if images are artificially extended as is done in \code{\link{PreProc}}.}

\item{savepassthru}{Logical; should intermediate files be saved to directory? Can greatly increase size on disk but useful to see which objects are being filtered out.}
}
\value{
Data frame containing all flux filtered sources.
}
\description{
Filter sources in data frame based on axial ratio value.
}
