% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_r2_mcfadden}
\alias{get_r2_mcfadden}
\title{Calcuate the the adjusted McFadden's pseudo R-square}
\usage{
get_r2_mcfadden(embeds, y)
}
\arguments{
\item{embeds}{a n-by-q matrix, specify the embedding matrix.}

\item{y}{a n-length vector, specify the labels.}
}
\value{
return the adjusted McFadden's pseudo R-square.
}
\description{
Calcuate the the adjusted McFadden's pseudo R-square  between the  embeddings and the  labels
}
\details{
None
}
\references{
McFadden, D. (1987). Regression-based specification tests for the multinomial logit model. Journal of econometrics, 34(1-2), 63-82.
}
