% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_Cutter.R
\name{Group_Cutter}
\alias{Group_Cutter}
\title{Group_Cutter}
\usage{
Group_Cutter(
  RA_DEC,
  image_directory,
  asteroids = NULL,
  frames = NULL,
  segim = NULL,
  groupim = NULL
)
}
\arguments{
\item{RA_DEC}{String; Celestial Right Ascension and Declination of Input Frame separated by underscore (RA_Dec).}

\item{image_directory}{String; Location of input fits files. Default behaviour is to check working directory for suitable fits images.}

\item{asteroids}{Numeric Data Frame; Optional. Data frame containing astronomical bodies to produce cutouts of. If not supplied will look for appropriate file in working directory.}

\item{frames}{List; Optional. List containing three pixel-matched astronomical images. Expected order within list is (g,r,i), returned by \code{\link{ProFAst::Pre_Proc}}. If not supplied will look for appropriate file in working directory and perform pixel matching. Can greatly increase execution time if images are sufficiently large.}

\item{segim}{Integer Matrix; Optional. Matrix containing object segmentation masks for an image. If not supplied will look for appropriate file in working directory.}

\item{groupim}{Integer Matrix; Optional. Matrix containing group segmentation masks for an image. If not supplied will look for appropriate file in working directory.}
}
\description{
Postage stamp maker. Produces cutouts of fits images, with segmentation mask data overlaid.
}
