\name{pbmc3k_subset}  
\alias{pbmc3k_subset}  
\title{A Seurat object including scRNA-seq PBMC dataset}  
\description{  
  This data is a subset of  PBMC3k scRNA-seq data in SeuratData package.
}  
\usage{  
  data(pbmc3k_subset)  
}  
\format{  
  A Seurat object, including count matrix, and manual annotation.
}  
\source{  
  The data is from the scRNA-seq sequencing platform.
}  
\references{  
  None
}  
\examples{  
  # Show  examples of how to use the dataset.  
  data(pbmc3k_subset)  
  library(Seurat)
  pbmc3k_subset
}  
\keyword{datasets}
