% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre_Proc.R
\name{Pre_Proc}
\alias{Pre_Proc}
\title{PreProc}
\usage{
Pre_Proc(RA_DEC, image_directory)
}
\arguments{
\item{${1:RA_DEC}}{${2:Celestial Right Ascension and Declination of Input Frame separated by underscore (RA_Dec).}}

\item{${1:image_directory}}{${2:Location of input fits files.}}
}
\value{
List containing 3 pixel matched images
}
\description{
${1:Function that preprocesses desired multiband images. Used to resize all images to same size, returning pixel matched images.}
}
