% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSCMEB.R
\name{RuniSCMEB}
\alias{RuniSCMEB}
\title{Fit an iSC-MEB model using the embeddings from FAST}
\usage{
RuniSCMEB(PRECASTObj, ...)
}
\arguments{
\item{PRECASTObj}{a PRECASTObj object created by \code{\link{CreatePRECASTObject}}.}

\item{...}{other arguments passed to \code{\link{iscmeb_run}}.}
}
\value{
Return a revised PRECASTObj object with an added component \code{iSCMEB} in the slot \code{PRECASTObj@resList} (including the aligned embeddings, clusters and posterior probability matrix of clusters).
}
\description{
Fit an iSC-MEB model using the embeddings from FAST and the number of clusters obtained by Louvain.
}
