% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SelectHKgenes}
\alias{SelectHKgenes}
\title{Select housekeeping genes}
\usage{
SelectHKgenes(seuList, species = c("Human", "Mouse"), HK.number = 200)
}
\arguments{
\item{seuList}{an M-length  list consisting of Seurat object, include the information of expression matrix and spatial coordinates (named \code{row} and \code{col}) in the slot \code{meta.data}.}

\item{species}{a string, the species, one of 'Human' and 'Mouse'.}

\item{HK.number}{an optional integer, specify the number of housekeeping genes to be selected.}
}
\value{
Return a string vector of the selected gene names.
}
\description{
Select housekeeping genes for preparation of removing unwanted variations in expression matrices
}
