\name{top5_signatures}  
\alias{top5_signatures}  
\title{A data.frame object including top five signature genes in scRNA-seq PBMC dataset}  
\description{  
  This data is a data.frame object that includes top five signature genes in scRNA-seq PBMC dataset
}  
\usage{  
  data(top5_signatures)  
}  
\format{  
  A data.frame object, including signature genes, distance, and manual annotation.
}  
\source{  
  None
}  
\references{  
  None
}  
\examples{  
  # Show  examples of how to use the dataset.  
  data(top5_signatures)  
  head(top5_signatures)
}  
\keyword{datasets}
