% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N100_Filter.R
\name{N100_Filter}
\alias{N100_Filter}
\title{N100_filter}
\usage{
N100_Filter(RA_DEC, N100_lower = 150, N100_upper = 2250, filtered_data = NULL)
}
\arguments{
\item{RA_DEC}{Celestial Right Ascension and Declination of Input Frame separated by underscore (RA_Dec).}

\item{N100_lower}{Small pixel count cutoff. Objects with fewer pixels than the threshold are assumed to be errors and excluded.}

\item{N100_upper}{Large pixel count cutoff. Objects with more pixels than the threshold are assumed to be errors and excluded.}

\item{filtered_data}{List; Filtered astronometric data. If not supplied will look for appropriate file in working directory.}
}
\value{
Data frame containing all N100 filtered sources.
}
\description{
Filter sources in data frame based on N100 value (pixel size).
}
