% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProFAst.R
\name{ProFAst}
\alias{ProFAst}
\title{ProFAst}
\usage{
ProFAst(
  RA_DEC,
  image_directory,
  skycut = 0.6,
  pixcut = 15,
  smooth = TRUE,
  sigma = 2,
  reltol = -10,
  tolerance = 1,
  ext = 7,
  flux_value = 1,
  edge_buffer = 0.001,
  axrat_value = 0.35,
  N100_lower = 150,
  N100_upper = 2250
)
}
\arguments{
\item{${1:RA_DEC}}{${2:Celestial Right Ascension and Declination of Input Frame separated by underscore (RA_Dec).}}

\item{${1:image_directory}}{${2:Location of input fits files.}}
}
\value{
Result of calculation
}
\description{
${1:Full execution of detection and filter runs. Extra input paramters are passed through to other function calls to adjust detection or filter processes.}
}
