% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSCMEB.R
\name{RunHarmonyLouvain}
\alias{RunHarmonyLouvain}
\title{Embedding alignment and clustering based on the embeddings from ProFAST}
\usage{
RunHarmonyLouvain(PRECASTObj, resolution = 0.5)
}
\arguments{
\item{PRECASTObj}{a PRECASTObj object created by \code{\link{CreatePRECASTObject}}.}

\item{resolution}{an optional real, the value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities.}
}
\value{
Return a revised \code{PRECASTObj} object with slot \code{PRECASTObj@resList} added by a \code{Harmony} compoonent (including the aligned embeddings and embeddings of batch effects) and a \code{Louvain} component (including the clusters).
}
\description{
Embedding alignment and clustering using the Harmony and Louvain based on the ebmeddings from ProFAST as well as determining the number of clusters.
}
